+++ assets/js/checkout-fields.js
@@ -70,8 +70,8 @@ jQuery(function ( $ ) {
 				enableFields( getBillingFields( country ) );
 			}
 
-			if ($( '#ebanx_billing_foreign_document_field' ).length > 0 && isEbanxMethodSelected() && country) {
-				let display = country === 'br' ? 'hide' : 'block';
+			if ( $( '#ebanx_billing_foreign_document_field' ).length > 0 && isEbanxMethodSelected() && country ) {
+				let display = country === 'br' ? 'none' : 'block';
 				$( '#ebanx_billing_foreign_document_field' ).css( { display: display } );
 			}